The story with the Code
Well this description can be learnt from the code, I will be uploading all my practice with life, my teachers teachings, my grandpas's blessings, the great sayings, to the one defined coding abilities and prove that I can become a coder easily, writing my daily experiences with family friends and life. This is the beginning to the greatest code of the history, which explains how I turn myself into coding and become a coder.